% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AUCell_Batch.R
\name{AUCell_batch}
\alias{AUCell_batch}
\title{Scores the SCE with AUCell}
\usage{
AUCell_batch(inp_data, genesets, num_batches = 100)
}
\arguments{
\item{inp_data}{SCE object}

\item{genesets}{geneset that you want test with the SCE object,amed list of character vectors, each consisting of a set of gene symbols}

\item{num_batches}{umber of batches to run AUCell for. More batches = fewer cells (observations)}
}
\value{
Returns an either an nxp matrix (samples x scores)
}
\description{
A function Scores a data matrix with AUCell in batches. Idea is to limit memory consumption when
scoring with AUCell
}
\examples{
expr=load_expression("mf_patient02.csv")
pathway_scores = AUCell_batch(counts(expr), genesets = c(hallmarks), num_batches=20)
# Access pathway_scores
pathway_scores


}
