% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSingleCellFile.R
\name{load_expression}
\alias{load_expression}
\title{Function that load dataframe Files}
\usage{
load_expression(exprs_file)
}
\arguments{
\item{expr_file}{A file that contains the cell information}
}
\value{
Returns single cell experiment object
}
\description{
A function that read dataframe that contain information about the cells
then convert it to a single cell experiment function
}
\examples{
expr<-load_expression(Patient_data)

# Access sce value
expr
}
