% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Patient_data}
\alias{Patient_data}
\title{SingleCell data from one myelofibrosis patients}
\format{
A matrix with columns:
\describe{
\item{cols}{Name of genes }
\item{Rows}{Barcodes of celss of Myelofibrosis Patients}
}
}
\source{
Bethan Psaila, et. al. Single-Cell Analyses Reveal Megakaryocyte-Biased Hematopoiesis in Myelofibrosis and Identify Mutant Clone-Specific Targets, Molecular Cell, Volume 78, Issue 3. 2020. Pages 477-492.e8. ISSN 1097-276. Web. https://www.sciencedirect.com/science/ article/pii/ S1097276520302343
}
\usage{
Patient_data
}
\description{
Patients data comes from Psaila.et al 2016 paper
}
\examples{
\dontrun{
 Patient_data
}
}
\keyword{datasets}
