% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPathwayScores.R
\name{PlotPathwayScores}
\alias{PlotPathwayScores}
\title{Plot pathway scores for given rows of cell}
\usage{
PlotPathwayScores(pathway_scores, start, end, pathway_names)
}
\arguments{
\item{pathway_scores}{Result get from AUCell_batch which contain the rank based on AUCell
of every cells in a geneset.The rows contains the barcodes of each cell.}

\item{start}{number of rows that you want start with}

\item{end}{number of rows that you want end}

\item{pathway_names}{the name of the pathway you want to compare with}
}
\value{
Returns a plot with y axis with cell barcodes and their expression level
of certain pathways
}
\description{
A function that plot and illustrate the enrichment of a given pathway for given cells
}
\examples{
expr=load_expression("mf_patient02.csv")
pathway_scores = AUCell_batch(counts(expr), genesets = c(hallmarks), num_batches=20)
Plot <- PlotPathwayScores(pathway_scores,1,3,HALLMARK_CHOLESTEROL_HOMEOSTASIS)
# Access plot
plot


}
