% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunRASC.R
\name{runRASC}
\alias{runRASC}
\title{Launch Shiny App for RASC}
\usage{
runRASC()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for RASC.
The shiny app permit to plot a histogram that showed comparison between certain
number of rows for a given pathway
}
\examples{
\dontrun{

RASC::runRASC()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
